ajustes e criando arquivo Docker

+++ src/Shared/MyBookRental.Communication/Requests/RequestReturnBookRentalJson.cs
@@ -2,6 +2,6 @@
 {
     public class RequestReturnBookRentalJson
     {
-        public long RentalId { get; set; } 
+        public int RentalId { get; set; } 
     }
 }
